requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Bot Framework
botbuilder-core>=4.14.0
botbuilder-integration-aiohttp>=4.14.0

# Database
supabase>=2.0.0

# HTTP
httpx>=0.25.0
aiohttp>=3.9.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Utils
python-dotenv>=1.0.0
structlog>=23.2.0
